+++ src/CarouselPaging.tsx
@@ -1,19 +1,18 @@
 /** @jsx createElement */
 import { createElement, FC } from 'react';
 import getModifierClass from './utils/getModifierClass';
+import { BemProps } from './types';
 
-export interface Props {
+export type Props = BemProps & {
 	current: number;
 	itemCount: number;
 	setCurrent: (idx: number) => void;
-	bem?: string;
-	modifier?: string | Array<string>;
 	texts?: {
 		next: string;
 		prev: string;
 		unit?: string;
 	};
-}
+};
 
 const defaultTexts: Props['texts'] = {
 	next: 'Next',